feat:Contributor Page:Move logic to Page component
 - Isolated ContributorCard component
 - Moved Data fetch operations to ContributorPage.
