+++ src/pages/ContributorPage.jsx
@@ -1,17 +1,18 @@
 import React, { useEffect, useState } from "react"
 import process from "process"
+import ContributorCard from "@/components/ContributorCard"
 
-function Contributors({ owner, repo }) {
+function ContriButorsPage() {
   const [contributors, setContributors] = useState([])
   const [loading, setLoading] = useState(true)
 
   useEffect(() => {
     getContributorData()
-  }, [owner, repo])
+  }, [])
 
   const getContributorData = async () => {
     setLoading(true) // set loading to true when th e fetch request is initiated
-    await fetch(`https://api.github.com/repos/${owner}/${repo}/contributors?per_page=100`, {
+    await fetch(`https://api.github.com/repos/priyankarpal/ProjectsHut/contributors?per_page=100`, {
       headers: {
         Authorization: ` ${process.env.GITHUB_TOKEN}`,
       },
@@ -26,23 +27,23 @@ function Contributors({ owner, repo }) {
         setLoading(false) // set loading to false when an error occurs
       })
   }
+
   return (
-    <section className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-8 justify-items-center items-center ">
+    <main className=" my-8 min-h-[65.75vh] max-w-6xl w-11/12 mx-auto sm:min-h-[73vh] lg:min-h-[78vh]">
+      <h1 className=" text-[1.5rem] md:text-[3rem] pb-5 font-bold  text-center">
+        Our <span className="text-primary">Contributors </span>
+      </h1>
       {loading ? (
         <div className="flex justify-center items-center h-full col-span-10">
           <span className="text-primary text-xl font-semibold">Loading...</span>
         </div>
       ) : (
-        contributors?.map((contributor) => (
-          <div className="min-w-max p-2" key={contributor.id}>
-            <a href={contributor.html_url} target="_blank">
-              <img className="h-16 mb-3 rounded-full " src={contributor.avatar_url} alt={contributor.login} />
-            </a>
-          </div>
-        ))
+        <section className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-8 justify-items-center items-center ">
+          <ContributorCard contributors={contributors} />
+        </section>
       )}
-    </section>
+    </main>
   )
 }
 
-export default Contributors
+export default ContriButorsPage
